refactor(game-engine): extract handleLeap/handleMove/handleDodge + handleDumpOffChoose (S27.8.12)
Slice de refactor S27.8 qui exploite la cassure de cycle d'import obtenue
via S27.8.11 (`getLegalMoves` extrait dans son propre module). Les 3
handlers `handleLeap`, `handleMove`, `handleDodge` etaient jusqu'ici
prisonniers de `actions.ts` car ils appellent `getLegalMoves` pour
valider la legalite du move. Maintenant qu'ils peuvent importer depuis
`./legal-moves`, l'extraction devient triviale.

`actions/move-leap-dodge-handlers.ts` (292l) regroupe :
- `handleLeap` (~74l) : Pogo Stick / Leap, AG roll + ball pickup
- `handleMove` (~34l) : aiguille handleDodgeRoll vs handleNormalMove
- `handleDodge` (~118l) : esquive explicite + Shadowing + Break Tackle

`handleDumpOffChoose` (53l) rejoint `actions/block-action.ts` plutot
qu'un module standalone : il termine en relancant `handleBlock` avec
`skipDumpOff: true`, donc cohesion thematique forte avec le flux de
blocage / dump-off.

Aucune dependance vers `actions.ts` depuis ces nouveaux modules : tous
les helpers proviennent de `mechanics/*`, `core/*`, `actions/legal-moves`,
`actions/move-handlers`, `actions/ball-pickup`, `actions/failure-helpers`.

Cleanup d'imports orphelins dans `actions.ts` (9 retires) :
`hasSkill`, `samePos`, `applyRollFailure`, `handleBallPickup`,
`handleDodgeRoll`, `handleNormalMove`, `executeDumpOff`,
`getLeapModifier`, `performLeapRoll`.

`actions.ts` shrinks 896 -> 626 lignes (-270 net lines, cumulative
-2080 vs origin 2706 = ~77% reduction, **26 lignes from the DoD
target of <= 600**).

`block-action.ts` 448 -> 503 lignes (+53 nettes pour le nouveau handler ;
toujours sous le seuil 800 du DoD).

8 new smoke tests in `move-leap-dodge-handlers.test.ts` cover the
3 exported functions, the missing-playerId guards, and state
immutability.

- 4918/4918 game-engine tests pass (+8 new smoke tests)
- 1787/1787 server tests pass (no regression)
- typecheck + lint + build clean (0 warnings on modified files)
